Made the stereo stuff almost (kinda) work
set_note_stereo seems to work with only a slight issue but play_music_stereo would take a lot more troubleshooting, it does play in stereo but not different notes and with some interesting bugs.
Main does not have stereo active in order to keep other parts working. We will need to expand the number of mappable pin in order to have stereo sound with everything else.